test(react_compiler): move end-to-end tests to tests/, drop internal unit tests
The 15 end-to-end tests (parse -> compile -> codegen via the public
transform_source API) move to tests/transform.rs as integration tests. The
internal unit tests (function discovery, JsString, HIR Environment, codegen
hashing, prefilter) are removed — the crate is exercised through its public API.

oxc_codegen moves to [dev-dependencies]: the library builds the oxc AST but
never codegens it; only the example and integration tests print the program.
