+++ crates/oxc_react_compiler/src/react_compiler/entrypoint/program.rs
@@ -3352,131 +3352,3 @@ fn convert_renames(
         })
         .collect()
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_is_hook_name() {
-        assert!(is_hook_name("useState"));
-        assert!(is_hook_name("useEffect"));
-        assert!(is_hook_name("use0Something"));
-        assert!(!is_hook_name("use"));
-        assert!(!is_hook_name("useless")); // lowercase after use
-        assert!(!is_hook_name("foo"));
-        assert!(!is_hook_name(""));
-    }
-
-    #[test]
-    fn test_is_component_name() {
-        assert!(is_component_name("MyComponent"));
-        assert!(is_component_name("App"));
-        assert!(!is_component_name("myComponent"));
-        assert!(!is_component_name("app"));
-        assert!(!is_component_name(""));
-    }
-
-    #[test]
-    fn test_is_valid_identifier() {
-        assert!(is_valid_identifier("foo"));
-        assert!(is_valid_identifier("_bar"));
-        assert!(is_valid_identifier("$baz"));
-        assert!(is_valid_identifier("foo123"));
-        assert!(!is_valid_identifier(""));
-        assert!(!is_valid_identifier("123foo"));
-        assert!(!is_valid_identifier("foo bar"));
-    }
-
-    /// Build an `oxc::FormalParameters` of plain identifier params from names,
-    /// for the component-param validity tests.
-    fn ident_params<'a>(
-        ast: &oxc_ast::builder::AstBuilder<'a>,
-        names: &[&str],
-    ) -> oxc_ast::ast::FormalParameters<'a> {
-        use oxc_span::SPAN;
-        let mut items = ast.vec();
-        for name in names {
-            let name = ox_atom(ast, name);
-            let pattern = ast.binding_pattern_binding_identifier(SPAN, name);
-            items.push(ast.formal_parameter(
-                SPAN,
-                ast.vec(),
-                pattern,
-                None::<oxc_allocator::Box<oxc_ast::ast::TSTypeAnnotation>>,
-                None::<oxc_ast::ast::Expression>,
-                false,
-                None,
-                false,
-                false,
-            ));
-        }
-        ast.formal_parameters(
-            SPAN,
-            oxc_ast::ast::FormalParameterKind::FormalParameter,
-            items,
-            None::<oxc_allocator::Box<oxc_ast::ast::FormalParameterRest>>,
-        )
-    }
-
-    #[test]
-    fn test_is_valid_component_params_empty() {
-        let allocator = oxc_allocator::Allocator::default();
-        let ast = oxc_ast::builder::AstBuilder::new(&allocator);
-        assert!(is_valid_component_params(&ident_params(&ast, &[])));
-    }
-
-    #[test]
-    fn test_is_valid_component_params_one_identifier() {
-        let allocator = oxc_allocator::Allocator::default();
-        let ast = oxc_ast::builder::AstBuilder::new(&allocator);
-        assert!(is_valid_component_params(&ident_params(&ast, &["props"])));
-    }
-
-    #[test]
-    fn test_is_valid_component_params_too_many() {
-        let allocator = oxc_allocator::Allocator::default();
-        let ast = oxc_ast::builder::AstBuilder::new(&allocator);
-        assert!(!is_valid_component_params(&ident_params(&ast, &["a", "b", "c"])));
-    }
-
-    #[test]
-    fn test_is_valid_component_params_with_ref() {
-        let allocator = oxc_allocator::Allocator::default();
-        let ast = oxc_ast::builder::AstBuilder::new(&allocator);
-        assert!(is_valid_component_params(&ident_params(&ast, &["props", "ref"])));
-    }
-
-    #[test]
-    fn test_should_skip_compilation_no_import() {
-        let allocator = oxc_allocator::Allocator::default();
-        let parsed = oxc_parser::Parser::new(
-            &allocator,
-            "function Component() {}\n",
-            oxc_span::SourceType::tsx(),
-        )
-        .parse();
-        let options = PluginOptions {
-            should_compile: true,
-            enable_reanimated: false,
-            is_dev: false,
-            filename: None,
-            compilation_mode: "infer".to_string(),
-            panic_threshold: "none".to_string(),
-            target: super::super::plugin_options::CompilerTarget::Version("19".to_string()),
-            gating: None,
-            dynamic_gating: None,
-            no_emit: false,
-            output_mode: None,
-            eslint_suppression_rules: None,
-            flow_suppressions: true,
-            ignore_use_no_forget: false,
-            custom_opt_out_directives: None,
-            environment: EnvironmentConfig::default(),
-            source_code: None,
-            profiling: false,
-            debug: false,
-        };
-        assert!(!should_skip_compilation(&parsed.program, &options));
-    }
-}
