+++ crates/oxc_react_compiler/src/react_compiler_hir/environment.rs
@@ -1007,84 +1007,3 @@ pub fn is_react_like_name(name: &str) -> bool {
     }
     is_hook_name(name)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_is_hook_name() {
-        assert!(is_hook_name("useState"));
-        assert!(is_hook_name("useEffect"));
-        assert!(is_hook_name("useMyHook"));
-        assert!(is_hook_name("use3rdParty"));
-        assert!(!is_hook_name("use"));
-        assert!(!is_hook_name("used"));
-        assert!(!is_hook_name("useless"));
-        assert!(!is_hook_name("User"));
-        assert!(!is_hook_name("foo"));
-    }
-
-    #[test]
-    fn test_environment_has_globals() {
-        let env = Environment::new();
-        assert!(env.globals().contains_key("useState"));
-        assert!(env.globals().contains_key("useEffect"));
-        assert!(env.globals().contains_key("useRef"));
-        assert!(env.globals().contains_key("Math"));
-        assert!(env.globals().contains_key("console"));
-        assert!(env.globals().contains_key("Array"));
-        assert!(env.globals().contains_key("Object"));
-    }
-
-    #[test]
-    fn test_get_property_type_array() {
-        let mut env = Environment::new();
-        let array_type = Type::Object { shape_id: Some("BuiltInArray".to_string()) };
-        let map_type = env.get_property_type(&array_type, "map").unwrap();
-        assert!(map_type.is_some());
-        let push_type = env.get_property_type(&array_type, "push").unwrap();
-        assert!(push_type.is_some());
-        let nonexistent = env.get_property_type(&array_type, "nonExistentMethod").unwrap();
-        assert!(nonexistent.is_none());
-    }
-
-    #[test]
-    fn test_get_function_signature() {
-        let env = Environment::new();
-        let use_state_type = env.globals().get("useState").unwrap();
-        let sig = env.get_function_signature(use_state_type).unwrap();
-        assert!(sig.is_some());
-        let sig = sig.unwrap();
-        assert!(sig.hook_kind.is_some());
-        assert_eq!(sig.hook_kind.as_ref().unwrap(), &HookKind::UseState);
-    }
-
-    #[test]
-    fn test_get_global_declaration() {
-        let mut env = Environment::new();
-        // Global binding
-        let binding = NonLocalBinding::Global { name: "Math".to_string() };
-        let result = env.get_global_declaration(&binding, None).unwrap();
-        assert!(result.is_some());
-
-        // Import from react
-        let binding = NonLocalBinding::ImportSpecifier {
-            name: "useState".to_string(),
-            module: "react".to_string(),
-            imported: "useState".to_string(),
-        };
-        let result = env.get_global_declaration(&binding, None).unwrap();
-        assert!(result.is_some());
-
-        // Unknown global
-        let binding = NonLocalBinding::Global { name: "unknownThing".to_string() };
-        let result = env.get_global_declaration(&binding, None).unwrap();
-        assert!(result.is_none());
-
-        // Hook-like name gets default hook type
-        let binding = NonLocalBinding::Global { name: "useCustom".to_string() };
-        let result = env.get_global_declaration(&binding, None).unwrap();
-        assert!(result.is_some());
-    }
-}
