+++ crates/oxc_react_compiler/src/react_compiler_hir/mod.rs
@@ -1519,45 +1519,3 @@ pub fn is_plain_object_type(ty: &Type) -> bool {
 pub fn is_start_transition_type(ty: &Type) -> bool {
     matches!(ty, Type::Function { shape_id: Some(id), .. } if id == object_shape::BUILT_IN_START_TRANSITION_ID)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_format_js_number() {
-        // Scientific notation for large numbers (>= 1e21)
-        assert_eq!(format_js_number(1e21), "1e+21");
-        assert_eq!(format_js_number(1.5e21), "1.5e+21");
-        assert_eq!(format_js_number(2.18739127891275e22), "2.18739127891275e+22");
-        assert_eq!(format_js_number(1e100), "1e+100");
-        assert_eq!(format_js_number(-1e21), "-1e+21");
-        assert_eq!(format_js_number(-1e100), "-1e+100");
-
-        // Scientific notation for small numbers (< 1e-6)
-        assert_eq!(format_js_number(1e-7), "1e-7");
-        assert_eq!(format_js_number(5e-7), "5e-7");
-        assert_eq!(format_js_number(1.5e-8), "1.5e-8");
-        assert_eq!(format_js_number(-1.5e-8), "-1.5e-8");
-
-        // Non-scientific large numbers (< 1e21)
-        assert_eq!(format_js_number(1e20), "100000000000000000000");
-        assert_eq!(format_js_number(1e-6), "0.000001");
-
-        // Integers
-        assert_eq!(format_js_number(0.0), "0");
-        assert_eq!(format_js_number(-0.0), "0");
-        assert_eq!(format_js_number(1.0), "1");
-        assert_eq!(format_js_number(100.0), "100");
-
-        // Regular floats
-        assert_eq!(format_js_number(1.5), "1.5");
-        assert_eq!(format_js_number(0.5), "0.5");
-        assert_eq!(format_js_number(0.1), "0.1");
-
-        // Special values
-        assert_eq!(format_js_number(f64::NAN), "NaN");
-        assert_eq!(format_js_number(f64::INFINITY), "Infinity");
-        assert_eq!(format_js_number(f64::NEG_INFINITY), "-Infinity");
-    }
-}
